pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"


[project]
name = "webpage-extractor"
version = "0.1.0"
description = "Webpage content extractor to Markdown + images, FastAPI, SOLID"
authors = [{ name = "Your Name" }]
requires-python = ">=3.11"
dependencies = [
"fastapi>=0.112",
"uvicorn[standard]>=0.30",
"jinja2>=3.1",
"httpx>=0.27",
"trafilatura>=1.9",
"pillow>=10.3",
"pydantic>=2.8",
"python-multipart>=0.0.9",
"beautifulsoup4>=4.12"
]


[tool.setuptools.packages.find]
where = ["."]


[tool.pytest.ini_options]
pythonpath = ["."]